WiP: now having a full functional migration of refSO3 (with positive and negatif pfn_offset)

+++ agency/avz/include/types.h
@@ -34,7 +34,7 @@
 
 /* mandatory types */
 
-typedef unsigned int size_t;
+typedef unsigned long size_t;
 typedef int ssize_t;
 
 typedef signed int ptrdiff_t;
